requirements.txt
beautifulsoup4==4.6.3
celery==4.2.1
Django==2.0.6
django-compressor==2.2
django-mptt==0.9.0
django-mptt-admin==0.5.1
django-jet==1.0.7
django-admin-sortable2==0.6.21
djangorestframework==3.8.2
psycopg2-binary==2.7.5
django-ckeditor==5.5.0
django-debug-toolbar==1.9.1
Pillow==5.1.0
redis==3.3.8
requests==2.21.0
sentry-sdk==0.7.4
uWSGI==2.0.17.1
telegram-send==0.20
htmlmin==0.1.10
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
urllib3==1.23
django-cacheops==4.1
